+++ galaxykit/tasks.py
@@ -2,16 +2,6 @@ import re
 from time import sleep
 
 
-def pulp_href_to_id(href):
-    uuid_regex = "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
-
-    for section in href.split("/"):
-        if re.match(uuid_regex, section):
-            return section
-
-    return None
-
-
 def get_tasks(client, only_running=False):
     tasks_url = f"pulp/api/v3/tasks/?ordering=-pulp_created"
     if only_running:
